Use incremental source generator
