+++ osu.Framework.SourceGeneration/DependencyInjectionSourceGenerator.cs
@@ -1,38 +1,128 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Linq;
+using System.Threading;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using osu.Framework.SourceGeneration.Emitters;
 
 namespace osu.Framework.SourceGeneration
 {
     [Generator]
-    public class DependencyInjectionSourceGenerator : ISourceGenerator
+    public class DependencyInjectionSourceGenerator : IIncrementalGenerator
     {
         protected virtual bool AddUniqueNameSuffix => true;
 
-        public void Initialize(GeneratorInitializationContext context)
+        public void Initialize(IncrementalGeneratorInitializationContext context)
         {
-            context.RegisterForSyntaxNotifications(() => new SyntaxContextReceiver());
+            IncrementalValuesProvider<GeneratorClassCandidate> candidateClasses =
+                context.SyntaxProvider.CreateSyntaxProvider(selectClasses, extractCandidates)
+                       .Where(c => c != null);
+
+            IncrementalValueProvider<(Compilation Compilation, ImmutableArray<GeneratorClassCandidate> classes)> compilationAndClasses =
+                context.CompilationProvider.Combine(candidateClasses.Collect());
+
+            context.RegisterSourceOutput(compilationAndClasses, emit);
         }
 
-        public void Execute(GeneratorExecutionContext context)
+        private bool selectClasses(SyntaxNode syntaxNode, CancellationToken cancellationToken)
         {
-            if (context.SyntaxContextReceiver is not SyntaxContextReceiver receiver)
+            if (syntaxNode is not ClassDeclarationSyntax classSyntax)
+                return false;
+
+            if (classSyntax.AncestorsAndSelf().OfType<ClassDeclarationSyntax>().Any(c => !c.Modifiers.Any(SyntaxKind.PartialKeyword)))
+                return false;
+
+            if (classSyntax.BaseList == null && classSyntax.AttributeLists.Count == 0)
+                return false;
+
+            return true;
+        }
+
+        private GeneratorClassCandidate extractCandidates(GeneratorSyntaxContext context, CancellationToken cancellationToken)
+        {
+            ClassDeclarationSyntax classSyntax = (ClassDeclarationSyntax)context.Node;
+            INamedTypeSymbol? symbol = context.SemanticModel.GetDeclaredSymbol(classSyntax);
+
+            if (symbol == null)
+                return null!;
+
+            // Determine if the class is a candidate for the source generator.
+            if (!symbol.AllInterfaces.Any(SyntaxHelpers.IsIDependencyInjectionCandidateInterface))
+                return null!;
+
+            GeneratorClassCandidate candidate = new GeneratorClassCandidate(classSyntax, symbol);
+
+            // Process any [Cached] attributes on any interface on the class excluding base types.
+            foreach (var iFace in SyntaxHelpers.GetDeclaredInterfacesOnType(symbol))
+            {
+                // Add an entry if this interface has a cached attribute.
+                if (iFace.GetAttributes().Any(attrib => SyntaxHelpers.IsCachedAttribute(attrib.AttributeClass)))
+                    candidate.CachedInterfaces.Add(iFace);
+            }
+
+            // Process any [Cached] attributes on the class.
+            foreach (var attrib in enumerateAttributes(context.SemanticModel, classSyntax))
+            {
+                if (SyntaxHelpers.IsCachedAttribute(context.SemanticModel, attrib))
+                    candidate.CachedClasses.Add(new SyntaxWithSymbol(context, classSyntax));
+            }
+
+            // Process any attributes of members of the class.
+            foreach (var member in classSyntax.Members)
+            {
+                foreach (var attrib in enumerateAttributes(context.SemanticModel, member))
+                {
+                    if (SyntaxHelpers.IsBackgroundDependencyLoaderAttribute(context.SemanticModel, attrib))
+                        candidate.DependencyLoaderMemebers.Add(new SyntaxWithSymbol(context, member));
+
+                    if (member is not PropertyDeclarationSyntax && member is not FieldDeclarationSyntax)
+                        continue;
+
+                    if (SyntaxHelpers.IsResolvedAttribute(context.SemanticModel, attrib))
+                        candidate.ResolvedMembers.Add(new SyntaxWithSymbol(context, member));
+
+                    if (SyntaxHelpers.IsCachedAttribute(context.SemanticModel, attrib))
+                        candidate.CachedMembers.Add(new SyntaxWithSymbol(context, member));
+                }
+            }
+
+            return candidate;
+        }
+
+        private static IEnumerable<AttributeSyntax> enumerateAttributes(SemanticModel semanticModel, MemberDeclarationSyntax member)
+        {
+            return member.AttributeLists
+                         .SelectMany(attribList =>
+                             attribList.Attributes
+                                       .Where(attrib =>
+                                           SyntaxHelpers.IsBackgroundDependencyLoaderAttribute(semanticModel, attrib)
+                                           || SyntaxHelpers.IsResolvedAttribute(semanticModel, attrib)
+                                           || SyntaxHelpers.IsCachedAttribute(semanticModel, attrib)));
+        }
+
+        private void emit(SourceProductionContext context, (Compilation compilation, ImmutableArray<GeneratorClassCandidate> candidates) items)
+        {
+            if (items.candidates.IsDefaultOrEmpty)
                 return;
 
-            foreach (var kvp in receiver.CandidateClasses)
-            {
-                GeneratorClassCandidate classCandidate = kvp.Value;
+            IEnumerable<GeneratorClassCandidate> distinctCandidates = items.candidates.Distinct();
+            ImmutableHashSet<ISymbol?> allClasses = items.candidates.Select(c => c.Symbol).ToImmutableHashSet(SymbolEqualityComparer.Default);
 
+            foreach (var candidate in distinctCandidates)
+            {
                 // Fully qualified name, with generics replaced with friendly characters.
-                string typeName = classCandidate.Symbol.ToDisplayString(SymbolDisplayFormat.CSharpErrorMessageFormat)
-                                                .Replace('<', '{')
-                                                .Replace('>', '}');
+                string typeName = candidate.Symbol.ToDisplayString(SymbolDisplayFormat.CSharpErrorMessageFormat)
+                                           .Replace('<', '{')
+                                           .Replace('>', '}');
 
                 string filename = $"g_{typeName}_Dependencies.cs";
 
-                context.AddSource(filename, new DependenciesFileEmitter(context, receiver, classCandidate).Emit());
+                context.AddSource(filename, new DependenciesFileEmitter(candidate, items.compilation, allClasses).Emit());
             }
         }
     }
